We split the large PR for adding Graph.h and Graph.cpp to `nativert` into 3 smaller PRs:
1. **Add header file**
2. Add source file
3. Add test and build rules.

Torch Native Runtime RFC: https://github.com/pytorch/rfcs/pull/72

4 classes have been introduced: `Graph`, `Node`, `Value`, `Type`
- `Type` represents the kind of a `Value`
- `Value` represents a single symbolic value, it could be any kind that exists in `Type`. Values are inputs and outputs of a `Node`.
- `Node` represents a single unit of execution, typically a PyTorch op.
- `Graph` represents a model's computation graph, which is designed to facilitate transformation/analysis.

Differential Revision: [D75491860](https://our.internmc.facebook.com/intern/diff/D75491860/)

[ghstack-poisoned]

+++ torch/_inductor/codecache.py
@@ -1975,7 +1975,7 @@ class AotCodeCompiler:
                 )
                 wrapper_build_options.save_flags_to_json(compile_flags)
                 generated_files.append(compile_flags)
-                wrapper_builder.save_compile_cmd_to_cmake(cmake_path, device_type)
+                wrapper_builder.save_compile_cmd_to_cmake(cmake_path)
                 wrapper_builder.save_src_to_cmake(cmake_path, wrapper_path)
                 generated_files.append(cmake_path)
             else:
@@ -2134,8 +2134,8 @@ class AotCodeCompiler:
                     so_builder.save_src_to_cmake(cmake_path, consts_o)
 
                 if config.aot_inductor.multi_arch_kernel_binary:
-                    so_builder.save_kernel_asm_to_cmake(cmake_path, asm_files)
-                    generated_files.extend(asm_files)
+                    # TODO: support multi-arch when package_cpp_only
+                    pass
                 else:
                     obj_srcs = [*gpu_kernels_o, *cubins_o]
                     generated_files.extend(obj_srcs)
@@ -2146,6 +2146,8 @@ class AotCodeCompiler:
             else:
                 so_builder.build()
                 for o_file in obj_srcs:
+                    if o_file in gpu_kernels_o:
+                        continue
                     # Remove these as they are not needed anymore
                     os.remove(o_file)
